+++ src/content_publisher/app/oauth/credentials.py
@@ -4,8 +4,6 @@ import pickle
 import os
 from datetime import datetime
 
-import time
-
 from typing import Dict, Any, List, Optional, Callable
 
 from content_publisher.app.oauth.oauth_flow import OAuthError
@@ -16,8 +14,7 @@ logger = logging.getLogger(__name__)
 class Credentials:
     def __init__(self, data: Dict[str, Any]):
         self.__data = copy.deepcopy(data)
-        self.__data['birthday'] = time.time() - 30  # 30 seconds buffer
-        self.__data['expires_in'] = self._init_expiry()
+        self.__data['expires_at'] = self._init_expires_at()
 
     @property
     def data(self) -> Dict[str, Any]:
@@ -35,10 +32,6 @@ class Credentials:
     def scopes(self)  -> List[str]:
         return self.__data.get('scopes', [])
 
-    @property
-    def expires_in(self) -> Optional[float]:
-        return self.data.get('expires_in', None)
-
     def with_scopes(self, scopes: List[str]) -> 'Credentials':
         new_data = self.__data.copy()
         new_data['scopes'] = scopes
@@ -54,26 +47,37 @@ class Credentials:
             return True if not self.scopes else set(self.scopes).issubset(set(scopes))
         return self.is_refreshable()
 
-    def is_expired(self) -> bool:
-        return self.expires_in is not None and time.time() - self.__data['birthday'] >= self.expires_in
-
-    def _init_expiry(self) -> Optional[float]:
-        for key in ['expires_in', 'expiry', 'expires_at']:
+    def _init_expires_at(self) -> Optional[str]:
+        for key in ['expires_in', 'expiry', 'expires_at', 'expires']:
             val = self.__data.get(key)
             if val is None or val == '':
                 continue
             try:
-                return float(val)
+                expires_in = float(val) - 30  # 30 seconds buffer
+                return datetime.fromtimestamp(datetime.now().timestamp() + expires_in).isoformat()
             except Exception:
-                try:
-                    expiry = datetime.strptime(str(val), '%Y-%m-%dT%H:%M:%S')
-                    return datetime.now().timestamp() - expiry.timestamp()
-                except Exception:
-                    logger.warning(f"Invalid expires_in. {self}")
+                return str(val)
         return None
 
+    def is_expired(self, fallback: bool = False) -> bool:
+        if 'expires_at' not in self.__data or self.__data['expires_at'] is None:
+            return fallback
+        try:
+            sval = str(self.__data['expires_at'])
+            if '.' in sval:
+                expiry = datetime.fromisoformat(sval)
+            else:
+                expiry = datetime.strptime(sval, '%Y-%m-%dT%H:%M:%S')
+            return datetime.now() >= expiry
+        except Exception:
+            return fallback
+
     def __str__(self):
-        return f"{self.__class__.__name__}(expires_in={self.expires_in}, scopes={self.scopes})"
+        return (f"{self.__class__.__name__}"
+                f"(is_expired={self.is_expired()}, expires_at={self.__data.get('expires_at')}, "
+                f"access_token={None if self.access_token is None else '***'},"
+                f"refresh_token={None if self.refresh_token is None else '***'},"
+                f"scopes={self.scopes})")
 
 
 class CredentialsStore:
@@ -84,20 +88,35 @@ class CredentialsStore:
             logger.debug(f"Created directory {self.dir_path}")
 
     def load_or_fetch(self,
-                      fetch: Callable[[Optional[Credentials]], Credentials],
                       filename: str,
-                      scopes: list[str]) -> 'Credentials':
+                      fetch: Callable[[Optional[Credentials]], Credentials],
+                      scopes: list[str],
+                      is_valid: Callable[[Credentials], bool] = lambda credentials: True) -> 'Credentials':
         stored_creds = self.load(filename, scopes)
-        if stored_creds and not stored_creds.is_expired():
+        if stored_creds and stored_creds.is_expired() is False and is_valid(stored_creds):
             logger.debug(f"Using existing: {stored_creds}")
             return stored_creds
+        self.delete(filename)
         fresh_creds = fetch(stored_creds)
         if fresh_creds:
             if 'error' in fresh_creds.data.keys():
                 raise OAuthError(fresh_creds.data)
             logger.debug(f"Using newly fetched: {fresh_creds}")
             self.save(filename, fresh_creds.with_scopes(scopes))
-        return fresh_creds    
+        return fresh_creds
+
+    def delete(self, filename: str) -> bool:
+        filename = self._file_path(filename)
+        if not os.path.exists(filename):
+            logger.warning(f"Not found, file: {filename}.")
+            return False
+        try:
+            os.remove(filename)
+            logger.debug(f"Deleted: {filename}")
+            return True
+        except Exception as ex:
+            logger.debug(f"Failed to delete: {filename}. Reason: {ex}")
+            return False
 
     def load(self, filename: str, scopes: List[str]) -> Optional[Credentials]:
         filename = self._file_path(filename)
@@ -106,18 +125,19 @@ class CredentialsStore:
             return None
         try:
             # logger.debug(f"Loading credentials from: {filename}")
-            with open(filename, 'rb') as token:
-                creds_data = pickle.load(token)
-                creds = Credentials(creds_data if creds_data else {})
+            with open(filename, 'rb') as credentials_file:
+                creds_data = pickle.load(credentials_file)
+                credentials = Credentials(creds_data if creds_data else {})
+                logger.debug(f"Loaded {credentials} from: {filename}")
 
-            if creds.is_valid(scopes):
-                return creds
+            if credentials.is_valid(scopes):
+                return credentials
             else:
-                try:
-                    os.remove(filename)
-                    logger.debug(f"Deleted invalid/expired credentials file: {filename}")
-                except Exception as ex:
-                    logger.debug(f"Failed to delete invalid/expired credentials file: {filename}. Reason: {ex}")
+                deleted = self.delete(filename)
+                if deleted:
+                    logger.debug(f"Deleted invalid/expired {credentials} file: {filename}")
+                else:
+                    logger.debug(f"Failed to delete invalid/expired {credentials} file: {filename}")
                 return None
         except Exception as ex:
             logger.warning(f"Could not load credentials from: {filename}. Reason: {ex}")
@@ -130,12 +150,13 @@ class CredentialsStore:
             if not os.path.exists(dirname):
                 os.makedirs(dirname, exist_ok=True)
                 logger.debug(f"Created directory {dirname}")
-            # logger.debug(f"Saving credentials to: {filename}")
+            # logger.debug(f"Saving {credentials} to: {filename}")
             with open(filename, 'wb') as credentials_file:
                 pickle.dump(credentials.data, credentials_file)
+                logger.debug(f"Saved {credentials} to: {filename}")
             return True
         except Exception as ex:
-            logger.warning(f"Could not save credentials to: {filename}. Reason: {ex}")
+            logger.warning(f"Could not save {credentials} to: {filename}. Reason: {ex}")
             return False
 
     def _file_path(self, filename: str):