+++ src/content_publisher/app/tiktok/tiktok_oauth.py
@@ -50,16 +50,21 @@ class TikTokOAuth(OAuth):
 
         def fetch_credentials(credentials: Optional[Credentials]):
             if credentials and credentials.is_expired() and credentials.is_refreshable():
+                logger.debug(f"Refreshing: {credentials}")
                 token_data = self._refresh_access_token(credentials.refresh_token)
-                return Credentials(token_data).with_scopes(scopes) if token_data else None
+                credentials = Credentials(token_data).with_scopes(scopes) if token_data else None
+                logger.debug(f"Refreshed: {credentials}")
+                return credentials
 
             token_data = self._exchange_auth_code_for_access_token(get_auth_code(), verifier_params)
-            return Credentials(token_data).with_scopes(scopes)
+            credentials = Credentials(token_data).with_scopes(scopes)
+            logger.debug(f"Fetched: {credentials}")
+            return credentials
 
         if not credentials_file:
             return fetch_credentials(None)
 
-        return self.credentials_store.load_or_fetch(fetch_credentials, credentials_file, scopes)
+        return self.credentials_store.load_or_fetch(credentials_file, fetch_credentials, scopes)
 
     def _build_auth_url(self, scopes: list[str], additional_params: Optional[dict[str, Any]] = None) -> str:
         params = {