+++ src/content_publisher/main.py
@@ -6,7 +6,6 @@ from app.app import App
 from app.content_publisher import Content
 from app.run_arg import RunArg
 
-
 if __name__ == "__main__":
     run_args = RunArg.of_sys_argv()
     logging.basicConfig(
@@ -15,17 +14,27 @@ if __name__ == "__main__":
     )
 
     platforms = run_args.get(RunArg.PLATFORMS)
+    print(f"Publishing to platforms: `{platforms}`")
     dir_path = run_args.get(RunArg.DIR)
     text_title = run_args.get(RunArg.TEXT_TITLE)
     media_orientation = run_args.get(RunArg.MEDIA_ORIENTATION)
     language_code = run_args.get(RunArg.LANGUAGE_CODE)
     tags = run_args.get(RunArg.TAGS)
 
-    content = Content.of_dir(dir_path, text_title, media_orientation, language_code, tags)
+    content = Content.of_dir(dir_path, title=text_title, media_orientation=media_orientation,
+                             language_code=language_code, tags=tags)
 
     configs = {
-        "facebook": { "credentials_scopes": ['business_management', 'pages_show_list'] },
+        "youtube": {
+            "dry_run": True,
+            "credentials_filename": "youtube-expired.pickle"
+        },
+        "facebook": {
+            "dry_run": True,
+            "credentials_scopes": ['business_management', 'pages_show_list']
+        },
         "tiktok": {
+            "dry_run": True,
             "callback_path": '/callback',
             # TODO - Remove this post_info, when we are able to post to TikTok
             #  with privacy PUBLIC TO EVERYONE (i.e the default)