+++ src/test/app/oauth/credentials_test.py
@@ -1,4 +1,5 @@
 import time
+from datetime import datetime, timedelta
 
 import unittest
 
@@ -33,6 +34,17 @@ class CredentialsTest(unittest.TestCase):
         credentials = Credentials(data)
         self.assertTrue(credentials.is_expired())
 
+
+    def test_expired_given_now(self):
+        data = {
+            'access_token': 'test-access-token',
+            'refresh_token': 'test-refresh-token',
+            'expires_at': datetime.now().isoformat(),
+            'scopes': ['test-scope-1', 'test-scope-2']
+        }
+        credentials = Credentials(data)
+        self.assertTrue(credentials.is_expired())
+
     def test_expired_given_negative(self):
         data = {
             'access_token': 'test-access-token',
@@ -43,11 +55,31 @@ class CredentialsTest(unittest.TestCase):
         credentials = Credentials(data)
         self.assertTrue(credentials.is_expired())
 
-    def test_not_expired(self):
+    def test_expired_given_past(self):
+        data = {
+            'access_token': 'test-access-token',
+            'refresh_token': 'test-refresh-token',
+            'expires_at': datetime.now() - timedelta(hours=1),
+            'scopes': ['test-scope-1', 'test-scope-2']
+        }
+        credentials = Credentials(data)
+        self.assertTrue(credentials.is_expired())
+
+    def test_not_expired_given_positive(self):
         data = {
             'access_token': 'test-access-token',
             'refresh_token': 'test-refresh-token',
             'expires_in': 10000,
+            'scopes': ['test-scope-1', 'test-scope-2']
+        }
+        credentials = Credentials(data)
+        self.assertFalse(credentials.is_expired())
+
+    def test_not_expired_given_future(self):
+        data = {
+            'access_token': 'test-access-token',
+            'refresh_token': 'test-refresh-token',
+            'expires_at': datetime.now() + timedelta(hours=1),
             'scopes': ['test-scope-1', 'test-scope-2']
         }
         credentials = Credentials(data)